requirements.txt
feedparser==6.0.11
groq==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
